fixed getMeal function
- still missing the functionallity to display  a meal without any categories or ingredients
